requirements.txt
pytelegrambotapi
Django
polygon-api-client